% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHGVS.R
\name{new.hgvs.builder.c}
\alias{new.hgvs.builder.c}
\title{Coding Sequence HGVS Builder}
\usage{
new.hgvs.builder.c()
}
\value{
A \code{hgvs.builder.c} object with functions for building coding HGVS strings. 
  The individual functions return single-element character vectors containing these strings.
}
\description{
A constructor for a CDS (=coding sequence) HGVS builder object. The object contains a collection of functions
for building CDS HGVS strings.
The resulting object encapsulates the following functions:
\itemize{
	\item{substitution(pos,ancestral,variant,posOffset=0)} CDS substitution variants. 
   pos = position (integer); ancestral = ancestral nucleotide [ACGT]; 
   variant = variant nucleotide [ACGT]; posOffset = offset from the position when
   crossing exon-intron borders (integer, defaults to 0)
	\item{deletion(start,stop,startOffset=0,stopOffset=0)} CDS deletion. start = start position (integer);
   stop = stop position (integer); startOffset = offset from the start position when
   crossing exon-intron borders (integer, defaults to 0); stopOffset = offset from the 
   stop position when crossing exon-intron borders (integer, defaults to 0)
	\item{inversion(start,stop,startOffset=0,stopOffset=0)} CDS inversion. start = start position (integer);
   stop = stop position (integer); startOffset = offset from the start position when
   crossing exon-intron borders (integer, defaults to 0); stopOffset = offset from the 
   stop position when crossing exon-intron borders (integer, defaults to 0)
	\item{duplication(start,stop,startOffset=0,stopOffset=0)} CDS duplication. start = start position (integer);
   stop = stop position (integer); startOffset = offset from the start position when
   crossing exon-intron borders (integer, defaults to 0); stopOffset = offset from the 
   stop position when crossing exon-intron borders (integer, defaults to 0)
	\item{insertion(start,variant,startOffset=0)} CDS insertion. start = position immediately preceeding 
   the insertion (integer); seq = inserted nucleotide sequence [ACGT]+ ; 
   startOffset = offset from the start position when crossing exon-intron borders 
   (integer, defaults to 0)
	\item{delins(start,stop,variant,startOffset=0,stopOffset=0)} CDS deletion and insertion. start = start position (integer); 
   stop = stop position relative to the reference (integer); 
   seq = inserted nucleotide sequence [ACGT]+ ; startOffset = offset from the start position when
   crossing exon-intron borders (integer, defaults to 0); stopOffset = offset from the 
   stop position when crossing exon-intron borders (integer, defaults to 0)
 \item{cis(...)} Multi-variant phased in cis. Parameters are coding HGVS strings for the 
   corresponding single mutants
 \item{trans(...)} Multi-variant phased in trans. Parameters are coding HGVS strings for the 
   corresponding single mutants
 \item{nophase(...)} Multi-variant with unknown phasing. Parameters are coding HGVS strings for the 
   corresponding single mutants
}
}
\examples{
builder <- new.hgvs.builder.c()
string1 <- builder$substitution(123,"A","G",posOffset=2)
string2 <- builder$delins(123,129,"ATTG")
string3 <- with(builder,cis(substitution(123,"A","C"),substitution(231,"G","A")))
}
\keyword{HGVS}
\keyword{builder}
