% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHGVS.R
\name{new.hgvs.builder.g}
\alias{new.hgvs.builder.g}
\title{Genomic HGVS Builder}
\usage{
new.hgvs.builder.g()
}
\value{
A \code{hgvs.builder.g} object with functions for building genomic HGVS strings. 
  The individual functions return single-element character vectors containing these strings.
}
\description{
A constructor for a genomic-level HGVS builder object. The object contains a collection of functions
for building genomic HGVS strings.
}
\details{
The resulting object encapsulates the following functions:
\itemize{
	\item{substitution(pos,ancestral,variant)} Genomic substitution variants. 
   pos = position (integer); ancestral = ancestral nucleotide [ACGT]; 
   variant = variant nucleotide [ACGT]
	\item{deletion(start,stop)} Genomic deletion. start = start position (integer);
   stop = stop position (integer)
	\item{inversion(start,stop)} Genomic inversion. start = start position (integer);
   stop = stop position (integer)
	\item{duplication(start,stop)} Genomic duplication. start = start position (integer);
   stop = stop position (integer)
	\item{insertion(start,variant)} Genomic insertion. start = position immediately preceeding 
   the insertion (integer); seq = inserted nucleotide sequence [ACGT]+
	\item{delins(start,stop,variant)} Genomic deletion and insertion. start = start position (integer); 
   stop = stop position relative to the reference (integer); seq = inserted nucleotide sequence [ACGT]+
 \item{cis(...)} Multi-variant phased in cis. Parameters are genomic HGVS strings for the 
   corresponding single mutants
 \item{trans(...)} Multi-variant phased in trans. Parameters are genomic HGVS strings for the 
   corresponding single mutants
 \item{nophase(...)} Multi-variant with unknown phasing. Parameters are genomic HGVS strings for the 
   corresponding single mutants
}
}
\examples{
builder <- new.hgvs.builder.g()
string1 <- builder$substitution(123,"A","G")
string2 <- builder$delins(123,129,"ATTG")
string3 <- with(builder,cis(substitution(123,"A","C"),substitution(231,"G","A")))
}
\keyword{HGVS}
\keyword{builder}
