% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHGVS.R
\name{new.hgvs.builder.p}
\alias{new.hgvs.builder.p}
\title{Protein HGVS Builder}
\usage{
new.hgvs.builder.p(aacode = c(1, 3))
}
\value{
A \code{hgvs.builder.g} object with functions for building genomic HGVS strings. 
  The individual functions return single-element character vectors containing these strings.
}
\description{
A constructor for a protein-level HGVS builder object. The object contains a collection of functions
for building protein HGVS strings.
}
\details{
The resulting object encapsulates the following functions:
\itemize{
 \item{synonymous()} A synonymous variant. No parameters required.
 \item{synonymous(pos,ancestral)} Unofficial (yet frequently used) version of synonymous variant syntax. 
   pos = position (integer); ancestral = ancestral amino acid in one-letter or three-letter code.
	\item{substitution(pos,ancestral,variant)} AA substitution variants. 
   pos = position (integer); ancestral = ancestral amino acid in one-letter or three-letter code; 
   variant = variant amino acid in one-letter or three-letter code
	\item{deletion(startPos,startAA,endPos,endAA)} AA deletion. startPos = start position (integer);
   startAA = start amino acid in one-letter or three-letter code;
   endPos = stop position (integer); endAA = start amino acid in one-letter or three-letter code
	\item{duplication(startPos,startAA,endPos,endAA)} AA duplication. startPos = start position (integer);
   startAA = start amino acid in one-letter or three-letter code;
   endPos = stop position (integer); endAA = start amino acid in one-letter or three-letter code
	\item{insertion(leftPos,leftAA,rightAA,seq)} AA insertion. leftPos = position immediately preceeding 
   the insertion (integer); leftAA = corresponding amino acid in one-letter or three-letter code;
   rightAA = amino acid to the right of the insertion, in one-letter or three-letter code;
   seq = inserted amino acid sequence, given as a character vector containing the individual
   one-letter or three-letter amino acid codes.
	\item{delins(startPos,startAA,endPos,endAA,seq)} AA deletion and insertion. 
   startPos = start position (integer);
   startAA = start amino acid in one-letter or three-letter code;
   endPos = stop position (integer); endAA = start amino acid in one-letter or three-letter code;
   seq = inserted amino acid sequence, given as a character vector containing the individual
   one-letter or three-letter amino acid codes.
 \item{frameshift(startPos,startAA,variantAA=NA,newStop=NA)} Frameshift variant. 
   startPos = start position (integer);
   startAA = start amino acid in one-letter or three-letter code;
   variantAA = amino acid replacing the start position in the frameshift sequence, 
   given in one-letter or three-letter code, or \code{NA} to omit (default);
   newStop = the position of the nearest coding resulting from the frameshift, 
   or \code{NA} to omit (default).
 \item{cis(...)} Multi-variant phased in cis. Parameters are coding HGVS strings for the 
   corresponding single mutants. As phasing in trans would be nonsensical in a protein context,
   the \code{trans()} and \code{nophase()} methods are not provided here.
}
}
\examples{
builder <- new.hgvs.builder.g()
string1 <- builder$substitution(123,"R","K")
string2 <- builder$delins(123,"Arg",152,"Leu",c("Lys","Trp","Ser"))
string3 <- with(builder,cis(substitution(123,"R","K"),deletion(125,"S",152,"L")))
}
\keyword{HGVS}
\keyword{builder}
