% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateHGVS.R
\name{translateHGVS}
\alias{translateHGVS}
\title{Translate a single HGVS string from nucleotide to amino acid level}
\usage{
translateHGVS(
  hgvs,
  cdsSeq,
  builder = new.hgvs.builder.p(aacode = 3),
  cbuilder = new.hgvs.builder.c(),
  strictMode = TRUE
)
}
\arguments{
\item{hgvs}{a single cds-level HGVS string. May contain multiple mutations in-cis.}

\item{cdsSeq}{the DNA coding sequence}

\item{builder}{an optional protein HGVS builder object, if none is provided, a new one is created.}

\item{strictMode}{whether HGVS will be validated against the reference sequence before translation}
}
\value{
a named vector containing the following strings: hgvsp: the translated HGVS string; 
  codonChanges: A simple string expression of the involved codon changes; codonHGVS: A codon-wise
  segmented HGVS string; aaChanges: A simple amino acid change string; and aaChangeHGVS: A codon-wise
  segmented HGVS string.
}
\description{
This function takes a codon-level HGVS string and translates it to amino acid level.
It does so in two different ways: A joint view, that describes the overall effect of all mutations
on the protein, as well as a codon-wise segmented way that describes changes on individual 
codons/amino-acids separately, so that they are more easily separable for Marginal Frequency analysis
}
