% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHGVS.R
\name{parseHGVS}
\alias{parseHGVS}
\title{HGVS Parser}
\usage{
parseHGVS(strings, aacode = c(1, 3, NA))
}
\arguments{
\item{strings}{A character vector containing the HGVS strings}

\item{aacode}{allowed values: 1, 3, or NA. Determines whether 1-letter codes or 3-letter codes should be forced. NA uses input format.}
}
\value{
A \code{data.frame} with the following columns:
}
\description{
Parses HGVS strings
}
\examples{
parseHGVS("g.1318G>T")
}
\keyword{HGVS}
\keyword{parsing}
